add test for default config.
